simplify __main__
